refactor(ir): introduce EmitExprNode for the post-strata emit-time tree
The bag-of-fields tree shape `resolvedToSlotted` produces and
`emit_numeric` consumes is structurally identical to the MCP wire-
format ExprNode but lives at a different semantic layer:

  ExprNode      — MCP wire format. What `wire`, `set_param`,
                  `define_program` accept. Combinators, ADTs, and
                  decl-bearing ops never appear at this boundary.
  EmitExprNode  — post-strata, post-resolvedToSlotted. The closed
                  set of slot-indexed refs + arithmetic + array ops
                  emit_numeric dispatches over.

Today they share an object shape, so the change is a type-alias
rename: `compiler/ir/emit_node.ts:EmitExprNode = number | boolean |
EmitExprNode[] | { op: string; [k]: unknown }`. Three consumers
switched to import the new alias:

- compiler/ir/lower_to_exprnode.ts (resolvedToSlotted output)
- compiler/ir/compile_resolved.ts (the lowered tree)
- compiler/emit_numeric.ts (the emit walker)

This unblocks D4: `compiler/expr.ts:ExprNode` can now narrow to MCP
wire-format ops only without affecting the emit boundary. The
eventual §2.1 from-scratch port of emit_numeric to walk
`ResolvedExpr` directly would let us delete EmitExprNode + the
lowering walker entirely.

Tests: 814 pass, 1 skip, 1 local-only fail. No behavior change.
